Move response interpretation to own class MODPATRON-66
