+++ src/main/java/org/folio/integration/http/ResponseInterpreter.java
@@ -1,18 +1,17 @@
-package org.folio.rest.impl;
+package org.folio.integration.http;
 
 import static org.apache.commons.lang3.StringUtils.isBlank;
 
 import java.util.concurrent.CompletionException;
 
-import org.folio.integration.http.Response;
 import org.folio.patron.rest.exceptions.HttpException;
 
 import io.vertx.core.json.JsonObject;
 
-class LookupsUtils {
-  private LookupsUtils() {}
+public class ResponseInterpreter {
+  private ResponseInterpreter() {}
 
-  static JsonObject verifyAndExtractBody(Response response) {
+  public static JsonObject verifyAndExtractBody(Response response) {
     if (!response.isSuccess()) {
       throw new CompletionException(new HttpException(response.statusCode,
         response.body));
